image-analysis-app/sam2/requirements.txt
fastapi
uvicorn
pillow
numpy
sam2 @ git+https://github.com/facebookresearch/sam2.git@2b90b9f5ceec907a1c18123530e92e794ad901a4
